Initial commit of Sha1++
This is a command line tool that computes the SHA-1 hash of a file, using C++17
features to statically generate the iterations of the hash algorithm. SHA-1
says approximately, "For iterations 10-19, use values a,b,c. For 20-29, use
i,j,k. Etc." So rather than the code checking at runtime which iteration it is
on, or even iterating, the unrolled sequence of steps are directly executed.